+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
+## 0.6.2
+
+- Added cache for slide/image handles
+- Fixed Leica rounding bug
+
 ## 0.6.1
 
 - removed explicit logging timestamps
-
 
 ## 0.6.0